refactor: move things into components
