+++ src/pages/index.js
@@ -1,122 +1,20 @@
-import React, { useEffect } from 'react';
-import Prism from 'prismjs';
+import React from 'react';
 import { graphql, Link } from 'gatsby';
-import styled from 'styled-components';
 import Layout from '../components/layout';
 import PostItem from '../components/postItem';
 import SEO from '../components/seo';
 import PageHeading from '../components/pageHeading';
-import Card from '../components/card';
 import { Copy, HeadingSmall, HeadingLarge } from '../components/type';
 import Newsletter from '../components/newsletterSignup';
+import Tools from '../components/tools';
 
-import { colors, padding } from '../theme';
-import HorizontalList from '../components/horizontalList';
-
-const HorizontalStack = styled(HorizontalList)`
-  margin-top: ${props => props.mt};
-  justify-content: space-between;
-`;
-
-const StackItem = styled.li`
-  width: 40%;
-  min-width: 300px;
-  flex-grow: 1;
-  :not(:last-child) {
-    margin-right: ${padding.large};
-    margin-bottom: ${padding.xlarge};
-  }
-`;
-
-const LinkButton = styled(Link)`
-  margin-right: ${padding.small};
-  text-decoration: none;
-  padding: ${padding.xsmall} ${padding.small};
-  border-radius: ${padding.xsmall};
-  border: 2px solid ${colors.blue};
-`;
-
-const SecondaryButton = styled(LinkButton)`
-  :hover {
-    color: ${colors.navy} !important;
-    background: ${colors.blue};
-  }
-`;
-
-const SecondaryLink = styled.a`
-  margin-right: ${padding.small};
-  text-decoration: none;
-  padding: ${padding.xsmall} ${padding.small};
-  border-radius: ${padding.xsmall};
-  border: 2px solid ${colors.blue};
-  :hover {
-    color: ${colors.navy} !important;
-    background: ${colors.blue};
-  }
-`;
-
-const PrimaryButton = styled(LinkButton)`
-  border-color: ${colors.green};
-  color: ${colors.green} !important;
-  :hover {
-    color: ${colors.navy} !important;
-    background: ${colors.green};
-  }
-`;
-
-const eKeysCode = `const inKeys = [
-  {
-    keyTime: 1,
-    keyValue: 0,
-    easeOut: 66,
-  },{
-    keyTime: 2,
-    keyValue: 250,
-    easeIn: 90,
-  }
-];
-
-eKeys.animate(inKeys, time);
-`;
-
-const eBoxCode = `const myBox =
-  eBox.createBox({
-    size: [200, 200],
-    position: [0, 0],
-    anchor: 'center',
-  });
-
-myBox.setScale(
-  [scaleIn, 100], 
-  'topLeft’
-);
-
-myBox.getPath();
-`;
-
-const CodeBlock = styled.div`
-  margin-bottom: 0;
-  box-shadow: 0 0 0 2pt ${colors.navy};
-`;
-
-const CodeJS = ({ children }) => {
-  return (
-    <CodeBlock className="gatsby-highlight" data-language="javascript">
-      <pre className="language-javascript">
-        <code className="language-javascript">{children}</code>
-      </pre>
-    </CodeBlock>
-  );
-};
+import { padding } from '../theme';
+import { HorizontalStack, StackItem } from '../components/stack';
+import LinkButton from '../components/linkButton';
 
 const BlogIndex = ({ data, location }) => {
   const siteTitle = data.site.siteMetadata.title;
   const posts = data.allMdx.edges;
-
-  useEffect(() => {
-    // call the highlightAll() function to style our code blocks
-    Prism.highlightAll();
-  });
 
   return (
     <Layout location={location} title={siteTitle}>
@@ -134,7 +32,9 @@ const BlogIndex = ({ data, location }) => {
             Master expressions, build templates, and improve your workflow by
             applying a developer mindset to motion design.
           </Copy>
-          <PrimaryButton to="/blog">Read the blog</PrimaryButton>
+          <LinkButton isLocal={true} type="primary" to="/blog">
+            Read the blog
+          </LinkButton>
         </StackItem>
         <StackItem>
           <h2>
@@ -144,61 +44,12 @@ const BlogIndex = ({ data, location }) => {
             We’ll free up your team by turning your most used graphics into
             templates your editors will love.
           </Copy>
-          <PrimaryButton to="/contact">Contact us</PrimaryButton>
+          <LinkButton isLocal={true} type="primary" to="/contact">
+            Contact us
+          </LinkButton>
         </StackItem>
       </HorizontalStack>
-      <h2>
-        <HeadingSmall>After Effects Tools</HeadingSmall>
-      </h2>
-      <Card>
-        <HorizontalStack mt="0">
-          <StackItem>
-            <h3>
-              <HeadingLarge>eKeys</HeadingLarge>
-            </h3>
-            <Copy>
-              Create keyframes within expressions - with full control over
-              easing.
-            </Copy>
-            <Copy>
-              {' '}
-              eKeys gives you the control you need to create dynamic templates,
-              with the smooth animation your used to with keyframes.{' '}
-            </Copy>
-            <SecondaryButton to="/blog/how-to-animate-with-expressions">
-              Read intro to eKeys
-            </SecondaryButton>
-          </StackItem>
-          <StackItem>
-            <CodeJS>{eKeysCode}</CodeJS>
-          </StackItem>
-        </HorizontalStack>
-      </Card>
-      <Card>
-        <HorizontalStack mt="0">
-          <StackItem>
-            <h3>
-              <HeadingLarge>eBox</HeadingLarge>
-            </h3>
-            <Copy>
-              Easily create, position and scale rectangles from any anchor
-              point.
-            </Copy>
-            <SecondaryLink href="https://github.com/motiondeveloper/ebox">
-              Get eBox
-            </SecondaryLink>
-          </StackItem>
-          <StackItem>
-            <CodeJS>{eBoxCode}</CodeJS>
-          </StackItem>
-        </HorizontalStack>
-      </Card>
-      <Copy>
-        <a href="https://github.com/motiondeveloper">
-          View all tools on Github
-        </a>
-      </Copy>
-
+      <Tools />
       <h2>
         <HeadingSmall>Featured Blog Posts</HeadingSmall>
       </h2>
